:bulb: extracted types from App component
